% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{color_level}
\alias{color_level}
\title{Adds/sets a specific level color, i.e. to make an "Other" level a specific
color outside the palette.}
\usage{
color_level(palette, level = "Other", color = "grey85")
}
\arguments{
\item{palette}{the named vector representing the palette}

\item{level}{level name (if missing, will be added to end)}

\item{color}{color to assign to specified level}
}
\description{
Adds/sets a specific level color, i.e. to make an "Other" level a specific
color outside the palette.
}
