% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{named_palette}
\alias{named_palette}
\title{Creates a named vector that can be used as in scale_*_manual}
\usage{
named_palette(levels, colors)
}
\arguments{
\item{levels}{the levels of a palette}

\item{colors}{the colors that go to each level. Must be at least as long as `levels`}
}
\description{
Creates a named vector that can be used as in scale_*_manual
}
