% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{tol_palette}
\alias{tol_palette}
\title{Paul Tol's qualitative color palettes for up to 21 discrete colors.
See https://personal.sron.nl/~pault/.}
\usage{
tol_palette(n)
}
\arguments{
\item{n}{the number of colors}
}
\description{
Paul Tol's qualitative color palettes for up to 21 discrete colors.
See https://personal.sron.nl/~pault/.
}
